CArtifactInstance moved to own file
